framework/requirements.txt
numpy
torchvision 
sklearn 
arff 

